gui: Factor out traceback widget
