+++ dangerzone/gui/main_window.py
@@ -14,19 +14,16 @@ from dangerzone.updater.releases import EmptyReport, ErrorReport, ReleaseReport
 if typing.TYPE_CHECKING:
     from PySide2 import QtCore, QtGui, QtSvg, QtWidgets
     from PySide2.QtCore import Qt
-    from PySide2.QtGui import QTextCursor
-    from PySide2.QtWidgets import QAction, QTextEdit
+    from PySide2.QtWidgets import QAction
 else:
     try:
         from PySide6 import QtCore, QtGui, QtSvg, QtWidgets
         from PySide6.QtCore import Qt
-        from PySide6.QtGui import QAction, QTextCursor
-        from PySide6.QtWidgets import QTextEdit
+        from PySide6.QtGui import QAction
     except ImportError:
         from PySide2 import QtCore, QtGui, QtSvg, QtWidgets
         from PySide2.QtCore import Qt
-        from PySide2.QtGui import QTextCursor
-        from PySide2.QtWidgets import QAction, QTextEdit
+        from PySide2.QtGui import QAction
 
 from .. import errors
 from ..document import SAFE_EXTENSION, Document
@@ -41,6 +38,7 @@ from ..updater import (
 )
 from ..util import format_exception, get_resource_path, get_version
 from .logic import Alert, CollapsibleBox, DangerzoneGui, UpdateDialog
+from .widgets import TracebackWidget
 
 log = logging.getLogger(__name__)
 
@@ -489,39 +487,6 @@ class WaitingWidget(QtWidgets.QWidget):
 
     def __init__(self) -> None:
         super(WaitingWidget, self).__init__()
-
-
-class TracebackWidget(QTextEdit):
-    """Reusable component to present tracebacks to the user.
-
-    By default, the widget is initialized but does not appear.
-    You need to call `.set_content("traceback")` on it so the
-    traceback is displayed.
-    """
-
-    def __init__(self) -> None:
-        super(TracebackWidget, self).__init__()
-        # Error
-        self.setReadOnly(True)
-        self.setVisible(False)
-        self.setProperty("style", "traceback")
-        # Enable copying
-        self.setTextInteractionFlags(Qt.TextSelectableByMouse)
-
-        self.current_output = ""
-
-    def set_content(self, error: Optional[str] = None) -> None:
-        if error:
-            self.setPlainText(error)
-            self.setVisible(True)
-
-    def process_output(self, line: str) -> None:
-        self.setVisible(True)
-        self.current_output += line
-        self.setText(self.current_output)
-        cursor = self.textCursor()
-        cursor.movePosition(QTextCursor.MoveOperation.End)
-        self.setTextCursor(cursor)
 
 
 class WaitingWidgetContainer(WaitingWidget):
